Restructure base64URL module
