--- encoding/base64url.json
@@ -1,12 +1,0 @@
-{
-  "name": "base64url",
-  "description": "Provides a way to encode strings and bytes from and into Base64URL.",
-  "signature": "{\n  decode: (text: string, context?: DecodeContext) => string;\n  decodeBytes: (\n    text: string,\n    context?: {\n      atob: (byteString: string) => string;\n      TextDecoder: new (encoding: string) => {\n        decode: (input?: Uint8Array) => string;\n      };\n    }\n  ) => number[];\n  encode: (\n    text: string,\n    context?: {\n      btoa: (byteString: string) => string;\n      TextEncoder: new () => {\n        encode: (input?: string) => Uint8Array;\n      };\n    }\n  ) => string;\n  encodeBytes: (bytes: number[], context?: EncodeContext) => string;\n  fromByteString: (byteString: string) => number[];\n  toByteString: (bytes: number[]) => string;\n}",
-  "examples": [
-    {
-      "language": "javascript",
-      "content": "base64url(); // ⇒ TODO"
-    }
-  ],
-  "questions": ["TODO: List questions that may this function answers."]
-}
