+++ encoding/base64url/DecodeContext.ts
@@ -1,0 +1,6 @@
+export interface DecodeContext {
+  atob: (byteString: string) => string;
+  TextDecoder: new (encoding: string) => {
+    decode: (input?: Uint8Array) => string;
+  };
+}
