+++ encoding/base64url/EncodeContext.ts
@@ -1,0 +1,6 @@
+export interface EncodeContext {
+  btoa: (byteString: string) => string;
+  TextEncoder: new () => {
+    encode: { (input?: string): Uint8Array };
+  };
+}
