+++ encoding/base64url/fromBase64.ts
@@ -1,0 +1,2 @@
+export default (base64: string) =>
+  base64.replace(/\+/g, "-").replace(/\//g, "_");
