+++ encoding/base64url/toBase64.ts
@@ -1,0 +1,2 @@
+export default (base64Url: string) =>
+  base64Url.replace(/-/g, "+").replace(/_/g, "/");
