+++ encoding/byteString/from.ts
@@ -1,0 +1,2 @@
+export default (byteString: string): number[] =>
+  [...byteString].map(_ => _.codePointAt(0) || 0);
