+++ encoding/byteString/to.ts
@@ -1,0 +1,2 @@
+export default (bytes: number[]) =>
+  bytes.map(_ => String.fromCharCode(_)).join("");
