--- encoding/index.js
@@ -1,5 +1,0 @@
-import base64url from "./base64url.js";
-
-export { base64url };
-
-export default { base64url };
