--- encoding/index.ts
@@ -1,5 +1,0 @@
-import base64url from "./base64url";
-
-export { base64url };
-
-export default { base64url };
